requirements.txt
markdown==3.6
setuptools==69.0.3
pyyaml==6.0.1